refactor: 로그인을 위한 UI 컴포넌트 단순화
